+++ test/rpc/test_deps.py
@@ -1,3 +1,5 @@
+import pytest
+
 from .util import (
     get_querier,
     ProjectDefinition,
@@ -70,7 +72,8 @@ def deps_with_packages(packages, bad_packages, project_dir, profiles_dir, schema
         querier.is_result(querier.async_wait(tok1))
 
 
-def test_rpc_deps_packages(project_root, profiles_root, postgres_profile, unique_schema):
+@pytest.mark.supported('postgres')
+def test_rpc_deps_packages(project_root, profiles_root, dbt_profile, unique_schema):
     packages = [{
         'package': 'fishtown-analytics/dbt_utils',
         'version': '0.2.1',
@@ -82,7 +85,8 @@ def test_rpc_deps_packages(project_root, profiles_root, postgres_profile, unique
     deps_with_packages(packages, bad_packages, project_root, profiles_root, unique_schema)
 
 
-def test_rpc_deps_git(project_root, profiles_root, postgres_profile, unique_schema):
+@pytest.mark.supported('postgres')
+def test_rpc_deps_git(project_root, profiles_root, dbt_profile, unique_schema):
     packages = [{
         'git': 'https://github.com/fishtown-analytics/dbt-utils.git',
         'revision': '0.2.1'