+++ test/rpc/test_run.py
@@ -1,3 +1,4 @@
+import pytest
 from .util import (
     assert_has_threads,
     get_querier,
@@ -5,8 +6,9 @@ from .util import (
 )
 
 
+@pytest.mark.supported('postgres')
 def test_rpc_run_threads(
-    project_root, profiles_root, postgres_profile, unique_schema
+    project_root, profiles_root, dbt_profile, unique_schema
 ):
     project = ProjectDefinition(
         models={'my_model.sql': 'select 1 as id'}
@@ -28,8 +30,9 @@ def test_rpc_run_threads(
         assert_has_threads(results, 7)
 
 
+@pytest.mark.supported('postgres')
 def test_rpc_run_vars(
-    project_root, profiles_root, postgres_profile, unique_schema
+    project_root, profiles_root, dbt_profile, unique_schema
 ):
     project = ProjectDefinition(
         models={
@@ -50,8 +53,9 @@ def test_rpc_run_vars(
         assert results['results'][0]['node']['compiled_sql'] == 'select 100 as id'
 
 
+@pytest.mark.supported('postgres')
 def test_rpc_run_vars_compiled(
-    project_root, profiles_root, postgres_profile, unique_schema
+    project_root, profiles_root, dbt_profile, unique_schema
 ):
     project = ProjectDefinition(
         models={